Remove Unexplained Addition to run_tests.py
Not sure what this bit about Haystack is in this script. Bob appears
to have set this up in his stuff but this edit is present in many of
my commits. Perhaps an old merge conflict that was never resolved?

Removing it to see if this solves the build failure problems.